vision_cursor/requirements.txt
opencv-python==4.8.0.74
mediapipe==0.10.5
PyQt5==5.15.9
pyautogui==0.9.54
SpeechRecognition==3.10.0
openai-whisper==20231117
numpy==1.24.3
dlib==19.24.2
pillow==10.0.0 